+++ lang/en/openbook.php
@@ -17,6 +17,7 @@ return [
         'loading' => 'Loading more posts...',
         'end' => 'There are no more posts to show.',
         'error' => 'Could not load more posts. Try reloading the page.',
+        'next' => 'More posts',
     ],
 
     'nav' => [